pyproject.toml
[tool.poetry]
name = "pinger"
version = "0.1.0"
description = ""
authors = ["iSplasher <isplasher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
